+++ src/screens/Home/index.tsx
@@ -7,13 +7,13 @@ import DailyQuestHeader from '@screens/Home/components/DailyQuestHeader';
 import DailyProgress from '@screens/Home/components/DailyProgress';
 import apiClient from '@apis/client';
 import Quest from '@type/Quest';
-import useUserDataQuery from '@hooks/useUserDataQuery';
-import useRegisteredQuestsQuery from '@hooks/useRegisteredQuestsQuery';
+import useUserDataQuery from '@hooks/queries/useUserDataQuery';
+import useRegisteredQuestsQuery from '@hooks/queries/useRegisteredQuestsQuery';
 import {
   useDeleteQuestMutation,
   useCompleteQuestMutation,
   useUpdateQuestsMutation,
-} from '@hooks/useQuestMutation';
+} from '@hooks/mutations/useQuestMutation';
 import ModalLayout from '@components/ModalLayout';
 import QuestModalContent from '@screens/Home/components/QuestModal/QuestModalContent';
 import SpacedView from '@components/common/SpacedView';
